Fix the centering algorithm so that the nroff and txt output files are more consistent.
Fix #129 - Left shift artwork that is greater than 69 characters wide and steal space from the left margin
Fix #151 & 194 which deal with how figures are layout
Fix #132 - if the artwork has an alignment - then it overrides the figure's version for the purpose of the artwork itself
Fix #213 - suppres-title kills the title decoration (i.e. Figure 1:) which matches v1 behavior.