Set of fixes that deal with xref in documents.
Set of fixes that deal with references.

Fix #209 - We now use the anchor rather than the generated bullet as the id of the reference element.
Fix #210 - The html did not have the same check for symrefs when sorting references that the text version did.  Copy it over so they both only sort if symrefs is yes
Fix #149 and #170- anchors on t elements in a section were referencable, but no those in lists.  They are now referencable
We now generate a warning when we get a target in an xref that we have not created an indexable reference for.  This basically gives us an internal error check.
We now generate a warning when a reference is created that is not targeted by an xref in the document.